// src/types.ts
export interface InlineAsyncData {
  type: string;
  pos: number;
  content: string;
  id: string;
}